[GUIS-10] Android: revise pwr log data to int.

+++ emgimuservice/src/main/java/org/sralab/emgimu/logging/FirebaseStreamLogger.java
@@ -199,7 +199,7 @@ public class FirebaseStreamLogger extends Observable {
         addJson(gson.toJson(msg));
     }
 
-    public void addPwrSample(long time, long sensor_timestamp, int sensor_counter, double [] data) {
+    public void addPwrSample(long time, long sensor_timestamp, int sensor_counter, int [] data) {
         Gson gson = new Gson();
         EmgPwrMessage msg = new EmgPwrMessage(mDeviceMac, time, sensor_timestamp, sensor_counter, data);
         addJson(gson.toJson(msg));
