+++ emgimuservice/src/main/java/org/sralab/emgimu/streaming/NetworkStreaming.java
@@ -163,7 +163,7 @@ public class NetworkStreaming {
 
     public void streamEmgPwr(BluetoothDevice dev,
                                 long time,
-                                double [] data) {
+                                int [] data) {
 
         Gson gson = new Gson();
         EmgPwrMessage msg = new EmgPwrMessage(dev.getAddress(), time, 0, 0, data);
